Change Layout of GUI
Change Layout of JFrame to CardLayout in order to switch JPanels.

Added additional JPanel to import existing test. This currently has
nothing on it, but it will have some buttons to make a file selector
dialog.

The other panel currently has some filler text areas and labels.

There is logic to flip to the next panel.