Add Custom Checks for Input Validation, Add Scroll Pane to Question Description Field
Added custom validateData method that would set a flag if the data
entered on a certain page was invalid. The method has to check to see
if the current panel index equals the index of the panel which the
elements reside on.

The scroll pane was achieved by adding a JScrollPane, setting vertical
scroll to as needed, and adding the JTextArea to it.