Revamp Question Adding System, Create External Dialogue
Question adding system revised so that a JList on a JScrollPane is used 
to actually display a list of questions available. Buttons to add, edit,
delete questions are provided and add/edit/delete on the list in
graphical form and in the backend.

The external dialog is a copy of what the original question creation
panels looked like. The YAML was just copied over.

Additionally, documented methods for the upcoming code check.