Copy Some Objects to Client Side, Fix Client Polling of IPs, Start Communication Between Client/Server
The IP polling problem was solved by fudging with the IOException
catch and finally clauses.

The Server was equipped with SSLServerSockets->ServerSockets while the
Client was equipped with Sockets. The model is 1 server, many clients.

Communication between the two was started on the Client side all the way
up to the receiving of Test data, which will be completed soon.